third_party/WebKit/LayoutTests/http/tests/push_messaging/has_permission_default.html

<!DOCTYPE html>
<html>
<head>
<title>Push API: hasPermission resolves with default when no permission is set</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            // If running manually, clear permissions before running this test.
            if (self.testRunner)
                testRunner.clearPushMessagingPermissions();
            return serviceWorkerRegistration.pushManager.hasPermission();
        })
        .then(function(permissionStatus) {
            assert_equals(permissionStatus, 'default');
            return service_worker_unregister_and_done(test, workerScope);
        })
        .catch(unreached_rejection(test));
}, 'Push API: hasPermission resolves with default when permission is not set');
</script>
</body>
</html>
